Создает контроллер Board и переносит туда всю логику компонентов внутри блока board
